refactors onboarding views
